+++ scout/server/blueprints/variants/forms.py
@@ -4,7 +4,7 @@ import logging
 
 from flask_wtf import FlaskForm
 from wtforms import (BooleanField, DecimalField, Field, TextField, SelectMultipleField,
-                     HiddenField, IntegerField, SubmitField, validators)
+                     SelectField, HiddenField, StringField, IntegerField, SubmitField, validators)
 from wtforms.widgets import TextInput
 from flask_wtf.file import FileField
 
@@ -48,14 +48,10 @@ class BetterDecimalField(DecimalField):
                 self.data = None
                 raise ValueError(self.gettext('Not a valid decimal value'))
 
-
-class FiltersForm(FlaskForm):
-    """Base FiltersForm for SNVs"""
+class VariantFiltersForm(FlaskForm):
     variant_type = HiddenField(default='clinical')
     gene_panels = SelectMultipleField(choices=[])
     hgnc_symbols = TagListField('HGNC Symbols/Ids (case sensitive)')
-
-    symbol_file = FileField('Symbol File')
 
     region_annotations = SelectMultipleField(choices=REGION_ANNOTATIONS)
     functional_annotations = SelectMultipleField(choices=FUNC_ANNOTATIONS)
@@ -64,10 +60,22 @@ class FiltersForm(FlaskForm):
     cadd_score = BetterDecimalField('CADD', places=2)
     cadd_inclusive = BooleanField('CADD inclusive')
     clinsig = SelectMultipleField('CLINSIG', choices=CLINSIG_OPTIONS)
+
+    gnomad_frequency = BetterDecimalField('gnomadAF', places=2)
+
+    filters = SelectField(choices=[])
+    filter_display_name = StringField(default='')
+    save_filter = SubmitField(label='Save filter')
+    load_filter = SubmitField(label='Load filter')
+    delete_filter = SubmitField(label='Delete filter')
+
+class FiltersForm(VariantFiltersForm):
+    """Base FiltersForm for SNVs - extends VariantFiltersForm."""
+    symbol_file = FileField('Symbol File')
+
     clinsig_confident_always_returned = BooleanField('CLINSIG Confident')
     spidex_human = SelectMultipleField('SPIDEX', choices=SPIDEX_CHOICES)
 
-    gnomad_frequency = BetterDecimalField('gnomadAF', places=2)
     chrom = TextField('Chromosome', [validators.Optional()])
     start = IntegerField('Start position', [validators.Optional(), IntegerField])
     end = IntegerField('End position', [validators.Optional(), IntegerField])
@@ -77,25 +85,15 @@ class FiltersForm(FlaskForm):
     clinical_filter = SubmitField(label='Clinical filter')
     export = SubmitField(label='Filter and export')
 
-class CancerFiltersForm(FlaskForm):
-    """Base filters for CancerFiltersForm"""
-    variant_type = HiddenField(default='clinical')
-    gene_panels = SelectMultipleField(choices=[])
-    region_annotations = SelectMultipleField(choices=REGION_ANNOTATIONS)
-    functional_annotations = SelectMultipleField(choices=FUNC_ANNOTATIONS)
-    genetic_models = SelectMultipleField(choices=GENETIC_MODELS)
-    hgnc_symbols = TagListField('HGNC Symbols/Ids (case sensitive)')
-    cadd_score = BetterDecimalField('CADD', places=2)
-    cadd_inclusive = BooleanField('CADD inclusive')
-    clinsig = SelectMultipleField('CLINSIG', choices=CLINSIG_OPTIONS)
+class CancerFiltersForm(VariantFiltersForm):
+    """Base filters for CancerFiltersForm - extends VariantsFiltersForm"""
     depth = IntegerField('Depth >')
     alt_count = IntegerField('Min alt count >')
     control_frequency = BetterDecimalField('Control freq. <', places=2)
     mvl_tag = BooleanField('In Managed Variant List')
-    gnomad_frequency = BetterDecimalField('gnomadAF', places=2)
 
 class StrFiltersForm(FlaskForm):
-    """docstring for CancerFiltersForm"""
+    """Placeholder filters form for CancerFiltersForm (**unused**)"""
     variant_type = HiddenField(default='clinical')
 
     chrom = TextField('Chromosome')